refactor: delete data_docs_site_anonymizer
